Make fromAddress in LockEvent and TransferTxArgs consistent
Previously fromAddress could either be the sender of the asset
or the txn initiator who is helping the asset owner deposit
through the wallet contract. This change makes it such that the
fromAddress will always be the initial asset owner.

Fixes issue 1) in https://github.com/Switcheo/switcheo-tradehub-eth/blob/master/audits/audit_2020_12_17.pdf.